Refactor with baseagent
