+++ lib/farscape.rb
@@ -1,2 +1,3 @@
 require 'plugins/plugins'
+require 'farscape/base_agent'
 require 'farscape/agent'
