+++ lib/farscape/agent.rb
@@ -3,6 +3,9 @@ require 'farscape/clients'
 
 module Farscape
   class Agent
+    
+    include BaseAgent
+    
     PROTOCOL = :http
 
     attr_reader :media_type
@@ -14,12 +17,6 @@ module Farscape
       @safe_mode = safe
       @plugin_hash = plugin_hash.empty? ? default_plugin_hash : plugin_hash
       handle_extensions
-    end
-
-    def handle_extensions
-      extensions = Plugins.extensions(enabled_plugins)
-      extensions = extensions[self.class.to_s.split(':')[-1].to_sym]
-      extensions.map { |cls| self.extend(cls) } if extensions
     end
 
     def representor
