+++ lib/farscape/representor.rb
@@ -4,6 +4,9 @@ require 'ostruct'
 
 module Farscape
   class SafeRepresentorAgent
+    
+    include BaseAgent
+    
     attr_reader :agent
     attr_reader :representor
     attr_reader :response
@@ -17,19 +20,9 @@ module Farscape
       @representor = deserialize(requested_media_type, response.body)
       handle_extensions
     end
-
-    def handle_extensions
-      extensions = Plugins.extensions(enabled_plugins)
-      extensions = extensions[self.class.to_s.split(':')[-1].to_sym]
-      extensions.map { |cls| self.extend(cls) } if extensions
-    end
     
     %w(using omitting).each do |meth|
       define_method(meth) { |name_or_type| self.class.new(@requested_media_type, @response, @agent.send(meth, name_or_type)) }
-    end
-
-    %w(disabled_plugins enabled_plugins plugins).each do |meth|
-      define_method(meth) { @agent.send(meth) }
     end
 
     def attributes
